move plugins to @xmpp/plugins
