+++ packages/component/README.md
@@ -2,7 +2,7 @@
 
 XMPP component for JavaScript
 
-Includes `@xmpp/component-core` and `@xmpp/plugin-reconnect`.
+Includes `@xmpp/component-core` and `@xmpp/plugins/reconnect`.
 
 ## Install
