--- packages/plugin-iq-caller/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@xmpp/plugin-iq-caller",
-  "description": "XMPP iq caller for JavaScript",
-  "repository": "github:node-xmpp/node-xmpp",
-  "homepage": "https://github.com/node-xmpp/node-xmpp/tree/master/packages/plugin-iq-caller",
-  "bugs": "http://github.com/node-xmpp/node-xmpp/issues",
-  "version": "0.0.0",
-  "license": "ISC",
-  "keywords": [
-    "XMPP",
-    "iq",
-    "caller",
-    "request"
-  ],
-  "dependencies": {
-    "@xmpp/plugin-router": "^0.0.0",
-    "@xmpp/xml": "^0.1.2"
-  }
-}
