--- packages/plugin-router/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@xmpp/plugin-router",
-  "description": "XMPP element router for JavaScript",
-  "repository": "github:node-xmpp/node-xmpp",
-  "homepage": "https://github.com/node-xmpp/node-xmpp/tree/master/packages/plugin-router",
-  "bugs": "http://github.com/node-xmpp/node-xmpp/issues",
-  "version": "0.0.0",
-  "license": "ISC",
-  "keywords": [
-    "XMPP",
-    "router"
-  ]
-}
